+++ src/DDTrace/Integrations/Guzzle/V5/GuzzleIntegration.php
@@ -3,13 +3,13 @@
 namespace DDTrace\Integrations\Guzzle\V5;
 
 use DDTrace\Configuration;
-use DDTrace\Formats;
-use DDTrace\Tags;
-use DDTrace\Span;
-use DDTrace\Types;
+use DDTrace\Format;
 use DDTrace\GlobalTracer;
 use DDTrace\Http\Urls;
 use DDTrace\Integrations\Integration;
+use DDTrace\Span;
+use DDTrace\Tag;
+use DDTrace\Type;
 use GuzzleHttp\Message\ResponseInterface;
 
 class GuzzleIntegration extends Integration