+++ src/DDTrace/Transport/Http.php
@@ -4,13 +4,12 @@ namespace DDTrace\Transport;
 
 use DDTrace\Configuration;
 use DDTrace\Encoder;
+use DDTrace\GlobalTracer;
 use DDTrace\Log\Logger;
 use DDTrace\Log\LoggerInterface;
 use DDTrace\Sampling\PrioritySampling;
 use DDTrace\Tracer;
 use DDTrace\Transport;
-use DDTrace\Version;
-use DDTrace\GlobalTracer;
 
 final class Http implements Transport
 {
@@ -54,7 +53,7 @@ final class Http implements Transport
         $this->setHeader('Datadog-Meta-Lang', 'php');
         $this->setHeader('Datadog-Meta-Lang-Version', \PHP_VERSION);
         $this->setHeader('Datadog-Meta-Lang-Interpreter', \PHP_SAPI);
-        $this->setHeader('Datadog-Meta-Tracer-Version', Version::VERSION);
+        $this->setHeader('Datadog-Meta-Tracer-Version', Tracer::VERSION);
     }
 
     /**