+++ src/DDTrace/Util/TryCatchFinally.php
@@ -2,9 +2,8 @@
 
 namespace DDTrace\Util;
 
+use DDTrace\Contracts\Span;
 use DDTrace\Scope;
-use DDTrace\Span;
-use DDTrace\SpanInterface;
 
 /**
  * PHP 5.4 compatible methods to workaround the missing try-catch-finally block.
@@ -26,7 +25,7 @@ class TryCatchFinally
     {
         $thrown = null;
         $result = null;
-        /** @var SpanInterface $span */
+        /** @var Span $span */
         $span = $scope->getSpan();
         try {
             $result = call_user_func_array([$instance, $method], $args);
@@ -60,7 +59,7 @@ class TryCatchFinally
     {
         $thrown = null;
         $result = null;
-        /** @var SpanInterface $span */
+        /** @var Span $span */
         $span = $scope->getSpan();
         try {
             $result = call_user_func_array($function, $args);